+++ activesupport/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   Add ActiveSupport::CurrentAttributes to provide a thread-isolated attributes singleton.
+    Primary use case is keeping all the per-request attributes easily available to the whole system.
+    
+    *DHH*
+
 *   Add support for versioned cache entries. This enables the cache stores to recycle cache keys, greatly saving
     on storage in cases with frequent churn. Works together with the separation of `#cache_key` and `#cache_version`
     in Active Record and its use in Action Pack's fragment caching.